merge new version

+++ artemis-server/src/main/java/org/apache/activemq/artemis/core/messagecounter/impl/MessageCounterHelper.java
@@ -38,7 +38,7 @@ public class MessageCounterHelper {
       DayCounterInfo[] infos = new DayCounterInfo[history.size()];
       for (int i = 0; i < infos.length; i++) {
          DayCounter dayCounter = history.get(i);
-         int[] counters = dayCounter.getCounters();
+         long[] counters = dayCounter.getCounters();
          GregorianCalendar date = dayCounter.getDate();
 
          DateFormat dateFormat = DateFormat.getDateInstance(DateFormat.SHORT);
@@ -141,8 +141,7 @@ public class MessageCounterHelper {
 
                if (value == -1) {
                   ret.append("<td></td>");
-               }
-               else {
+               } else {
                   ret.append("<td>" + value + "</td>");
 
                   total += value;
@@ -172,8 +171,7 @@ public class MessageCounterHelper {
    private static String asDate(final long time) {
       if (time > 0) {
          return DateFormat.getDateTimeInstance(DateFormat.SHORT, DateFormat.MEDIUM).format(new Date(time));
-      }
-      else {
+      } else {
          return "-";
       }
    }
